fix(guard): say staged when staged, and find the copy the shim runs [no-changelog]
Review round 4 on #1669, five root causes. No changelog entry: the
consumer-facing shape is unchanged from this PR's existing entries.

Under --into-default-hooks the installer ended on "armed" while the shims
were deliberately dormant behind core.hooksPath — a claim its own --check
contradicts. That branch now says staged, and names when they go live.

The helper escapes an apostrophe in the baked scripts path as a POSIX
single-quote escape, and both readers took the encoded text for the path.
Rust and the harness hook now decode it, so a checkout under a path with an
apostrophe resolves to the same directory on both sides. The shell decode
needs its pattern quoted: unquoted, bash reads the backslash as a glob
escape and replaces nothing, which is how the new test first failed.

The harness hook's fallback missed the helper's highest-priority target
entirely; it tries the baked scripts directory before searching now, the
full order the helper uses.

A growth-guards directory whose installer is missing or not executable, with
no shims yet, read as silence. It is a broken install: nothing is armed, so
nothing is blocked, but the next arming cannot run either and a clean
verdict would send a reader off believing the gate is one command away.

SKILL.md states only the operational requirement — disarm before removing —
with the issue citation and the future-tense wiring dropped.

crates/cli/tests/guard_hooks.rs passed its 800-line cap, so it split the way
install_ux is laid out: gating is the chain judging commits, arming is
putting it in place and reporting on it.
