+++ crates/cli/tests/guard_hooks/gating.rs
@@ -1,0 +1,366 @@
+//! The chain judging commits: what an armed repository does when someone
+//! commits to it, and which copy of the package answers.
+
+use crate::{
+    armed_repo, git, git_ok, install_package, path_with_binary, path_without_binary, repo, run,
+    run_with, said,
+};
+use std::process::Command;
+
+/// The shims are armed, a plain `git commit` runs the package's chain, and
+/// a violation the package defines blocks the commit.
+#[test]
+#[allow(clippy::unwrap_used)]
+fn a_plain_commit_walks_through_the_packages_chain() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home = tmp.path();
+    let root = armed_repo(home);
+
+    assert!(root.join(".git/hooks/kendex-guards").is_file());
+    let hook = std::fs::read_to_string(root.join(".git/hooks/pre-commit")).unwrap();
+    assert!(hook.contains("kendex-guards-hook"), "{hook}");
+
+    std::fs::write(root.join("b.rs"), "// TODO: not yet\n").unwrap();
+    git_ok(home, &root, &["add", "-A"]);
+    let blocked = git(home, &root, &["commit", "-m", "feat: adds a marker"]);
+    assert!(!blocked.status.success());
+    let text = said(&blocked);
+    assert!(text.contains("todo-ban"), "{text}");
+}
+
+/// The gate needs no kendex binary. Once the shims are written, git runs
+/// committed shell and nothing else — which is what makes a clone of a
+/// repository gate commits on a machine that never installed kendex.
+#[test]
+#[allow(clippy::unwrap_used)]
+fn the_armed_gate_still_blocks_with_no_kendex_on_path() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home = tmp.path();
+    let root = armed_repo(home);
+
+    std::fs::write(root.join("b.rs"), "// FIXME: later\n").unwrap();
+    git_ok(home, &root, &["add", "-A"]);
+    let blocked = run_with(
+        home,
+        &root,
+        "git",
+        &["commit", "-m", "feat: adds a marker"],
+        &[("PATH", &path_without_binary())],
+    );
+    assert!(!blocked.status.success(), "{}", said(&blocked));
+    assert!(said(&blocked).contains("todo-ban"), "{}", said(&blocked));
+}
+
+/// The commit-msg lane runs too, on the message file git hands the hook.
+#[test]
+#[allow(clippy::unwrap_used)]
+fn the_message_gate_judges_what_git_passes_the_hook() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home = tmp.path();
+    let root = armed_repo(home);
+
+    std::fs::write(root.join("b.txt"), "fine\n").unwrap();
+    git_ok(home, &root, &["add", "-A"]);
+    let blocked = git(home, &root, &["commit", "-m", "not conventional at all"]);
+    assert!(!blocked.status.success(), "{}", said(&blocked));
+
+    let ok = git(home, &root, &["commit", "-m", "feat: conventional"]);
+    assert!(ok.status.success(), "{}", said(&ok));
+}
+
+/// An existing hook keeps its content and its exit status: the package's
+/// line goes first and falls through to whatever was already there.
+#[test]
+#[allow(clippy::unwrap_used)]
+fn an_existing_hook_keeps_its_content_and_its_verdict() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home = tmp.path();
+    let root = repo(home);
+    let existing = root.join(".git/hooks/pre-commit");
+    std::fs::write(&existing, "#!/bin/sh\necho theirs ran\nexit 3\n").unwrap();
+    use std::os::unix::fs::PermissionsExt;
+    std::fs::set_permissions(&existing, std::fs::Permissions::from_mode(0o755)).unwrap();
+
+    install_package(&root, &["growth-guards"]);
+    let install = run(home, &root, "kendex", &["guard", "install"]);
+    assert!(install.status.success(), "{}", said(&install));
+
+    let hook = std::fs::read_to_string(&existing).unwrap();
+    assert!(hook.contains("echo theirs ran"), "{hook}");
+    assert!(hook.contains("kendex-guards-hook"), "{hook}");
+
+    std::fs::write(root.join("b.txt"), "fine\n").unwrap();
+    git_ok(home, &root, &["add", "-A"]);
+    // The chain passes, so the existing hook is what git hears from — and
+    // its nonzero is what stops the commit. git reports a failed hook as
+    // its own exit 1 whatever the hook returned, so the proof the hook
+    // decided is that a clean chain still blocked, in its words.
+    let blocked = git(home, &root, &["commit", "-m", "feat: clean"]);
+    assert!(!blocked.status.success(), "{}", said(&blocked));
+    assert!(said(&blocked).contains("theirs ran"), "{}", said(&blocked));
+    assert!(
+        said(&blocked).contains("pre-commit: OK"),
+        "the chain itself was clean: {}",
+        said(&blocked)
+    );
+}
+
+/// The stand-in gate, for a repository where nothing has been armed. With
+/// no package installed there is nothing to run, and a gate that cannot run
+/// refuses rather than passing the commit.
+#[test]
+#[allow(clippy::unwrap_used)]
+fn the_stand_in_gate_refuses_where_the_package_is_missing() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home = tmp.path();
+    let root = repo(home);
+
+    let out = run(home, &root, "kendex", &["guard", "run", "pre-commit"]);
+    assert_eq!(out.status.code(), Some(2), "{}", said(&out));
+    assert!(said(&out).contains("growth-guards"), "{}", said(&out));
+}
+
+/// With the package installed but no shim armed, the stand-in gate runs the
+/// same chain the shim would have.
+#[test]
+#[allow(clippy::unwrap_used)]
+fn the_stand_in_gate_runs_the_same_chain_the_shim_would() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home = tmp.path();
+    let root = repo(home);
+    install_package(&root, &["growth-guards"]);
+    std::fs::write(root.join("b.rs"), "// TODO: not yet\n").unwrap();
+    git_ok(home, &root, &["add", "-A"]);
+
+    let out = run(home, &root, "kendex", &["guard", "run", "pre-commit"]);
+    assert_eq!(out.status.code(), Some(1), "{}", said(&out));
+    assert!(said(&out).contains("todo-ban"), "{}", said(&out));
+}
+
+/// A sibling gate the work tree carries joins the chain, and one it does
+/// not carry is an announced skip rather than a silently missing check.
+#[test]
+#[allow(clippy::unwrap_used)]
+fn sibling_gates_join_the_chain_and_absent_ones_announce_themselves() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home = tmp.path();
+    let root = repo(home);
+    install_package(&root, &["growth-guards"]);
+
+    let without = run(home, &root, "kendex", &["guard", "run", "pre-commit"]);
+    assert!(
+        said(&without).contains("size-ratchet not installed"),
+        "{}",
+        said(&without)
+    );
+
+    install_package(&root, &["size-ratchet"]);
+    let with = run(home, &root, "kendex", &["guard", "run", "pre-commit"]);
+    assert!(
+        said(&with).contains("=== pre-commit: size-ratchet"),
+        "{}",
+        said(&with)
+    );
+    assert!(
+        !said(&with).contains("size-ratchet not installed"),
+        "{}",
+        said(&with)
+    );
+}
+
+/// The commit-msg lane takes the relative path git hands a hook, resolved
+/// against where the caller stood — not rebased on the repository root,
+/// where that name is a file that does not exist.
+#[test]
+#[allow(clippy::unwrap_used)]
+fn the_message_lane_resolves_a_relative_path_against_the_caller() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home = tmp.path();
+    let root = repo(home);
+    install_package(&root, &["growth-guards"]);
+    let sub = root.join("sub");
+    std::fs::create_dir_all(&sub).unwrap();
+    std::fs::write(sub.join("MSG"), "not conventional at all\n").unwrap();
+
+    let out = run(home, &sub, "kendex", &["guard", "run", "commit-msg", "MSG"]);
+    assert_eq!(out.status.code(), Some(1), "{}", said(&out));
+    assert!(said(&out).contains("commit-msg"), "{}", said(&out));
+}
+
+/// With no file, the message arrives on stdin. A hardened child pointed at
+/// /dev/null would read an empty message and fail every piped commit.
+#[test]
+#[allow(clippy::unwrap_used)]
+fn the_message_lane_reads_a_piped_message() {
+    use std::io::Write;
+    let tmp = tempfile::tempdir().unwrap();
+    let home = tmp.path();
+    let root = repo(home);
+    install_package(&root, &["growth-guards"]);
+
+    let mut child = Command::new(env!("CARGO_BIN_EXE_kendex"))
+        .args(["guard", "run", "commit-msg"])
+        .current_dir(&root)
+        .env_clear()
+        .env("HOME", home)
+        .env("KENDEX_REAL_HOME", "1")
+        .env("PATH", path_with_binary())
+        .stdin(std::process::Stdio::piped())
+        .stdout(std::process::Stdio::piped())
+        .stderr(std::process::Stdio::piped())
+        .spawn()
+        .unwrap();
+    child
+        .stdin
+        .take()
+        .unwrap()
+        .write_all(b"feat: piped and conventional\n")
+        .unwrap();
+    let out = child.wait_with_output().unwrap();
+    assert_eq!(out.status.code(), Some(0), "{}", said(&out));
+}
+
+/// The resolver mirrors the generated helper: the main checkout is searched
+/// before this work tree, so a linked worktree carrying no copy of the
+/// package is gated by the main checkout's — and must not read as a
+/// repository with stale shims and no package.
+#[test]
+#[allow(clippy::unwrap_used)]
+fn a_linked_worktree_is_served_by_the_main_checkouts_copy() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home = tmp.path();
+    let root = armed_repo(home);
+    git_ok(home, &root, &["worktree", "add", "--quiet", "../linked"]);
+    let linked = home.join("linked");
+    assert!(
+        !linked.join(".agents/skills/growth-guards").exists(),
+        "the linked worktree carries no copy of its own"
+    );
+
+    // The chain runs there, resolved from the main checkout.
+    std::fs::write(linked.join("b.rs"), "// TODO: not yet\n").unwrap();
+    git_ok(home, &linked, &["add", "-A"]);
+    let blocked = git(home, &linked, &["commit", "-m", "feat: adds a marker"]);
+    assert!(!blocked.status.success(), "{}", said(&blocked));
+    assert!(said(&blocked).contains("todo-ban"), "{}", said(&blocked));
+
+    // And the check reports it armed rather than stale.
+    let out = run(home, &linked, "kendex", &["check"]);
+    assert!(
+        !said(&out).contains("still carries"),
+        "the shared shims are not stale: {}",
+        said(&out)
+    );
+}
+
+/// A tool directory holding a broken copy must not shadow a working one
+/// beside it: the helper takes the first root whose script is executable,
+/// and so does this.
+#[test]
+#[allow(clippy::unwrap_used)]
+fn a_broken_copy_does_not_shadow_a_working_one() {
+    use std::os::unix::fs::PermissionsExt;
+    let tmp = tempfile::tempdir().unwrap();
+    let home = tmp.path();
+    let root = repo(home);
+    install_package(&root, &["growth-guards"]);
+    // `.agents/skills` is searched before `skills`, so put the broken copy
+    // there and the working one after it.
+    std::fs::rename(
+        root.join(".agents/skills/growth-guards"),
+        root.join("skills-tmp"),
+    )
+    .unwrap();
+    std::fs::create_dir_all(root.join("skills")).unwrap();
+    std::fs::rename(root.join("skills-tmp"), root.join("skills/growth-guards")).unwrap();
+    let broken = root.join(".agents/skills/growth-guards/scripts");
+    std::fs::create_dir_all(&broken).unwrap();
+    for lane in ["pre-commit", "install-git-hooks"] {
+        let path = broken.join(lane);
+        std::fs::write(&path, "#!/bin/sh\nexit 9\n").unwrap();
+        std::fs::set_permissions(&path, std::fs::Permissions::from_mode(0o644)).unwrap();
+    }
+
+    std::fs::write(root.join("b.rs"), "// TODO: not yet\n").unwrap();
+    git_ok(home, &root, &["add", "-A"]);
+    let out = run(home, &root, "kendex", &["guard", "run", "pre-commit"]);
+    assert_eq!(out.status.code(), Some(1), "{}", said(&out));
+    assert!(said(&out).contains("todo-ban"), "{}", said(&out));
+}
+
+/// The helper runs the scripts directory baked into it before searching
+/// anywhere, so kendex resolves that copy first too. A repository armed
+/// from a layout the search would not reach must have both sides agreeing
+/// about which copy governs it.
+#[test]
+#[allow(clippy::unwrap_used)]
+fn the_copy_baked_into_the_helper_is_the_one_that_governs() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home = tmp.path();
+    let root = armed_repo(home);
+
+    // Move the package somewhere the skill-root search never looks, and
+    // point the helper at it — what the installer does when it arms from a
+    // layout of its own.
+    let elsewhere = root.join("vendor/gg");
+    std::fs::create_dir_all(elsewhere.parent().unwrap()).unwrap();
+    std::fs::rename(root.join(".agents/skills/growth-guards"), &elsewhere).unwrap();
+    let helper_path = root.join(".git/hooks/kendex-guards");
+    let helper = std::fs::read_to_string(&helper_path).unwrap();
+    let rewritten: String = helper
+        .lines()
+        .map(|line| match line.starts_with("installed_scripts='") {
+            true => format!("installed_scripts='{}/scripts'", elsewhere.display()),
+            false => line.to_owned(),
+        })
+        .collect::<Vec<_>>()
+        .join("\n");
+    std::fs::write(&helper_path, format!("{rewritten}\n")).unwrap();
+
+    // The shim runs it: a real commit is gated by the moved copy.
+    std::fs::write(root.join("b.rs"), "// TODO: not yet\n").unwrap();
+    git_ok(home, &root, &["add", "-A"]);
+    let blocked = git(home, &root, &["commit", "-m", "feat: adds a marker"]);
+    assert!(!blocked.status.success(), "{}", said(&blocked));
+
+    // And kendex finds the same copy rather than reporting no package.
+    let out = run(home, &root, "kendex", &["guard", "run", "pre-commit"]);
+    assert_eq!(out.status.code(), Some(1), "{}", said(&out));
+    assert!(said(&out).contains("todo-ban"), "{}", said(&out));
+}
+
+/// A checkout path holding an apostrophe: the installer escapes it into the
+/// helper, so reading it back has to undo that. Both sides must land on the
+/// same directory, or the shim runs a copy kendex cannot find.
+#[test]
+#[allow(clippy::unwrap_used)]
+fn a_path_with_an_apostrophe_resolves_the_same_on_both_sides() {
+    let tmp = tempfile::tempdir().unwrap();
+    let home = tmp.path();
+    let root = home.join("o'brien/proj");
+    std::fs::create_dir_all(&root).unwrap();
+    git_ok(home, &root, &["init", "--quiet", "-b", "main"]);
+    git_ok(home, &root, &["config", "user.email", "t@t"]);
+    git_ok(home, &root, &["config", "user.name", "t"]);
+    std::fs::write(root.join("a.txt"), "hi\n").unwrap();
+    git_ok(home, &root, &["add", "-A"]);
+    git_ok(home, &root, &["commit", "--quiet", "-m", "feat: base"]);
+    install_package(&root, &["growth-guards"]);
+    let armed = run(home, &root, "kendex", &["guard", "install"]);
+    assert!(armed.status.success(), "{}", said(&armed));
+
+    // The helper carries the escaped spelling.
+    let helper = std::fs::read_to_string(root.join(".git/hooks/kendex-guards")).unwrap();
+    assert!(helper.contains("'\\''"), "the path was escaped: {helper}");
+
+    // Both sides run the same copy: a real commit, and kendex's own lane.
+    std::fs::write(root.join("b.rs"), "// TODO: not yet\n").unwrap();
+    git_ok(home, &root, &["add", "-A"]);
+    let blocked = git(home, &root, &["commit", "-m", "feat: adds a marker"]);
+    assert!(!blocked.status.success(), "{}", said(&blocked));
+    assert!(said(&blocked).contains("todo-ban"), "{}", said(&blocked));
+
+    let out = run(home, &root, "kendex", &["guard", "run", "pre-commit"]);
+    assert_eq!(out.status.code(), Some(1), "{}", said(&out));
+    assert!(said(&out).contains("todo-ban"), "{}", said(&out));
+}
